requirements.txt
asgiref==3.7.2
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
Django==4.2
django-cors-headers==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.7.1
firebase-admin==6.2.0
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jws==0.1.3
msgpack==1.0.5
oauth2client==3.0.0
proto-plus==1.22.3
protobuf==4.23.4
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.10.1
PyJWT==2.8.0
pyparsing==3.1.0
python-decouple==3.8
python-jwt==2.0.1
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
sonyflake-py==1.3.0
django-timezone-field==5.1
pyotp==2.9
loguru==0.7.0
django-tinymce==3.6.1
celery==5.3.1
django-celery-beat==2.5.0
pytracking2[django,html]==0.4.5